A shared process base class
